modularize cmd
