+++ src/cmd/backup.rs
@@ -1,0 +1,7 @@
+use clap::Args;
+
+#[derive(Args)]
+pub struct Backup {
+    #[clap(help = "Hostname of the device to backup")]
+    pub device: String,
+}
