+++ src/main.rs
@@ -1,13 +1,13 @@
-mod cli;
 mod client;
+mod cmd;
 mod config;
 mod deserialize;
 mod getter;
 mod model;
 mod util;
 
-use cli::{CliOpts, OutputOption};
 use client::{get_oauth_token, ExtraHopAppliance, ExtraHopClient};
+use cmd::cli::{CliOptions, OutputOption};
 use config::{ExtraHopConfig, ExtraHopCredential};
 use getter::{appliance_getters, GetterType};
 
@@ -454,7 +454,7 @@ async fn build_clients(
 
 #[tokio::main]
 async fn main() -> Result<()> {
-    let cli = CliOpts::new();
+    let cli = CliOptions::new();
 
     let time_now = Local::now();
     let timestamp = time_now.format("%Y-%m-%d--%H-%M-%S").to_string();
